[HeatMap] Store floats instead of doubles
This reduces the space to its half and doesn't lead to a precision loss
because Vector2 only stores floats anyway.

+++ src/KernelPanic/PathPlanning/BreadthFirstSearch.cs
@@ -23,7 +23,7 @@ namespace KernelPanic.PathPlanning
         {
             foreach (var node in Run(mGoalPoints))
             {
-                HeatMap.SetCost(node.Position, node.Cost);
+                HeatMap.SetCost(node.Position, (float) node.Cost);
             }
 
             VectorField.Update(HeatMap);
